labels/current/Label_Manager/requirements.txt
Pillow>=10.0.0
pytesseract>=0.3.10
ttkthemes>=3.2.2
python-dotenv>=1.0.0
